requirements.txt
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.97.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg2==2.9.6
pydantic==1.10.9
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
typing-extensions==4.6.3
uvicorn[standard]==0.22.0
pyhibp==4.2.0
python-jose[cryptography]
passlib[bcrypt]
pyotp==2.8.0
itsdangerous
requests
python-multipart